+++ src/test/java/com/ticketproject/webapp/controllers/AttendeeControllerTest.java
@@ -114,7 +114,7 @@ class AttendeeControllerTest
         EventHost eventHost,
         EventType eventType,
         RegistrationStatus registrationStatus,
-        int maxAttendees
+        long maxAttendees
     )
     {
         Event event = new Event.Builder()
